Feature/code revisions (#9)
* [VME-1621] Refactor to remove rails app patching to test different versions.

* [VME-1621] Refactor circleci config for DRY.

* [VME-1621] Fix for rubocop.

* [VME-1621] Remove stale paths from rubocop config.

* [VME-1621] Update README.

* [VME-1621] Localize change password page.

* [VME-1621] Fix test rake files to not abort when dep is missing.

In a dev instance you cannot run any other rake commands if all of our deps are not met. That's bad.

* [VME-1621] Refactor password_content_enforcement for flay.

* [VME-1621] Fix dummy app dev config to support local debugging without listen gem.

The 'listen' gem needs to be available if the config.file_watcher is configured. It doesn't make sense to install it for local testing so we just comment out the config.

* [VME-1621] Refactor password_frequent_reuse to use a scope instead of purging previous passwords.

Instead of purging old passwords we will just keep them and apply a scope on the association instead. That way a user can increase the number of previous passwords in the config later on but immediately reap its benefits.

* [VME-1621] Rename dppe_passwords_controller to passwords_with_policy_controller.

* [VME-1621] Refactor to remove LocaleTools lib.

* [VME-1621] Refactor to remove Math lib.

* [VME-1621] Refactor to remove Comparator lib.

* [VME-1621] Refactor to move fresh? and stale? methods to previous_password.

* [VME-1621] Rename insane_password? to inconsistent_password?

* [VME-1621] Add benchmark spec for validate_password_frequent_reuse.

* [VME-1621] Update circleci config to add ruby-2.5.0 build (in addition to 2.4.3).

* [VME-1621] Add benchmark spec for character_counter.